Created utilities file for functions without a clear organisational place yet.
